stun_rs::attributes

Module stun

STUN attributes defined for STUN protocol RFC8489

Modules

nonce_cookie

The Nonce Cookie used for Long-Term Credential Mechanism

Structs

AlternateServer

The alternate server represents an alternate transport address identifying a different STUN server that the STUN client should try.

ErrorCode

The ERROR-CODE attribute is used in error response messages.

MappedAddress

The MAPPED-ADDRESS attribute indicates a reflexive transport address of the client.

Nonce

The NONCE attribute may be present in requests and responses. It contains a sequence of qdtext or quoted-pair, which are defined in RFC3261. Note that this means that the NONCE attribute will not contain the actual surrounding quote characters.

PasswordAlgorithm

The PASSWORD-ALGORITHM attribute is present only in requests. It contains the algorithm that the server must use to derive a key from the long-term password.

PasswordAlgorithms

The PasswordAlgorithms attribute may be present in requests and responses. It contains the list of algorithms that the server can use to derive the long-term password.

Realm

The REALM attribute may be present in requests and responses. It contains text that meets the grammar for “realm-value” as described in RFC3261 but without the double quotes and their surrounding white space. That is, it is an unquoted realm-value (and is therefore a sequence of qdtext or quoted-pair).

Software

The Software attribute contains a textual description of the software being used by the agent sending the message. It is used by clients and servers. Its value SHOULD include manufacturer and version number. The attribute has no impact on operation of the protocol and serves only as a tool for diagnostic and debugging purposes.

UnknownAttributes

The UnknownAttributes is present only in an error response when the response code in the ERROR-CODE attribute is 420 (Unknown Attribute).

UserHash

The USER-HASH attribute is used as a replacement for the USER-NAME attribute when user name anonymity is supported.

UserName

The USER-NAME attribute is used for message integrity. It identifies the user name and password combination used in the message-integrity check. It MUST contain a UTF-8-encoded RFC3629 sequence of fewer than 509 bytes and MUST have been processed using the OpaqueString profile RFC8265

XorMappedAddress

The XorMappedAddress attribute is identical to the MappedAddress attribute, except that the reflexive transport address is obfuscated through the XOR function.

Enums

Fingerprint

The Fingerprint attribute MAY be present in all STUN messages. When present, the Fingerprint attribute MUST be the last attribute in the message and thus will appear after MessageIntegrity and MessageIntegritySha256.

MessageIntegrity

The MessageIntegrity attribute contains an HMAC-SHA1 RFC2104 of the STUN message. This attribute can be present in any STUN message type.

MessageIntegritySha256

The MessageIntegritySha256 attribute contains an HMAC-SHA256 RFC2104 of the STUN message. This attribute can be present in any STUN message type.